backend/requirements.txt
fastapi
uvicorn
spacy
openai-whisper
ffmpeg-python
